# Swagger/swagger.yaml
openapi: 3.0.0
servers:
    # Added by API Auto Mocking Plugin
    - description: SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/699853_1/BankingAPI/1.0.0
    - description: SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/MAHBAAN77/bankingApi/1.0.0
info:
    version: "1.0.0"
    title: Banking API
    description: API for managing banking operations, specifically for employee interactions with customer accounts and transactions.
    contact:
        email: you@your-company.com
    license:
        name: Apache 2.0
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
security:
    - JWTAuth: []
components:
    securitySchemes:
        JWTAuth:
            type: apiKey
            in: header
            name: Authorization
    schemas:
        CustomerAccount:
            type: object
            properties:
                accountId:
                    type: string
                    description: The unique identifier of the customer's account.
                customerId:
                    type: string
                    description: The unique identifier of the customer.
                accountType:
                    type: string
                    description: The type of the account (checking or savings).
                balance:
                    type: number
                    description: The current balance in the account.
                currency:
                    type: string
                    description: The currency of the account (ISO 4217 code).
                    enum:
                        - EUR
            required:
                - accountId
                - customerId
                - accountType
                - balance
                - currency
...
paths:
...
    /employees/customer-accounts:
        get:
            tags:
                - employees
            summary: View all customer accounts
            description: Retrieves a list of all customer accounts in the system.
            security:
                - JWTAuth: []
            responses:
                "200":
                    description: A list of customer accounts.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/CustomerAccount"
...